requirements.txt
Flask==3.1.2
openpyxl==3.1.5
PyPDF2==3.0.1
python-dotenv==1.1.1
python_jose==3.5.0
reportlab==4.4.3
requests==2.32.5
sendgrid==6.12.4
supabase==2.21.1
